+++ dwds/lib/src/injected/client.js
@@ -9898,7 +9898,7 @@
               rng[8] = rng[8] & 63 | 128;
               runId = A.UuidParsing_unparse(rng);
               $async$goto = 7;
-              return A._asyncAwait(manager.hotRestart$1$runId(runId), $async$handleWebSocketHotRestartRequest);
+              return A._asyncAwait(manager.hotRestart$2$reloadedSourcesPath$runId(A._asStringQ(init.G.$reloadedSourcesPath), runId), $async$handleWebSocketHotRestartRequest);
             case 7:
               // returning from await.
               A._sendResponse(clientSink, A.hot_restart_response_HotRestartResponse___new_tearOff$closure(), requestId, null, true, type$.HotRestartResponse);
@@ -22088,7 +22088,6 @@
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
       var t1, t2, value, $$v, _$result,
-        _s11_ = "BuildResult",
         result = new A.BuildResultBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
       for (t1 = type$.BuildStatus; iterator.moveNext$0();) {
@@ -22112,13 +22111,7 @@
         }
       }
       _$result = result._build_result$_$v;
-      if (_$result == null) {
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_build_result$_$this()._status, _s11_, "status", t1);
-        _$result = new A._$BuildResult(t2);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s11_, "status", t1);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.BuildResult);
-      return result._build_result$_$v = _$result;
+      return result._build_result$_$v = _$result == null ? new A._$BuildResult(A.BuiltValueNullFieldError_checkNotNull(result.get$_build_result$_$this()._status, "BuildResult", "status", t1)) : _$result;
     },
     deserialize$2(serializers, serialized) {
       return this.deserialize$3$specifiedType(serializers, serialized, B.FullType_null_List_empty_false);
@@ -22249,22 +22242,13 @@
       return _this;
     },
     _connect_request$_build$0() {
-      var t1, t2, t3, t4, _this = this,
+      var t1, _this = this,
         _s14_ = "ConnectRequest",
-        _s10_ = "instanceId",
-        _s14_0 = "entrypointPath",
         _$result = _this._connect_request$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._connect_request$_appId, _s14_, "appId", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._instanceId, _s14_, _s10_, t1);
-        t4 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._entrypointPath, _s14_, _s14_0, t1);
-        _$result = new A._$ConnectRequest(t2, t3, t4);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s14_, "appId", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s14_, _s10_, t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s14_, _s14_0, t1);
+        _$result = new A._$ConnectRequest(A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._connect_request$_appId, _s14_, "appId", t1), A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._instanceId, _s14_, "instanceId", t1), A.BuiltValueNullFieldError_checkNotNull(_this.get$_connect_request$_$this()._entrypointPath, _s14_, "entrypointPath", t1));
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ConnectRequest);
       return _this._connect_request$_$v = _$result;
     }
   };
@@ -22429,23 +22413,13 @@
       return _this;
     },
     _debug_event$_build$0() {
-      var t1, t2, t3, t4, t5, _this = this,
+      var t1, _this = this,
         _s10_ = "DebugEvent",
-        _s9_ = "eventData",
-        _s9_0 = "timestamp",
         _$result = _this._debug_event$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._debug_event$_kind, _s10_, "kind", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._eventData, _s10_, _s9_, t1);
-        t4 = type$.int;
-        t5 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._timestamp, _s10_, _s9_0, t4);
-        _$result = new A._$DebugEvent(t2, t3, t5);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s10_, "kind", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s10_, _s9_, t1);
-        A.BuiltValueNullFieldError_checkNotNull(t5, _s10_, _s9_0, t4);
+        _$result = new A._$DebugEvent(A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._debug_event$_kind, _s10_, "kind", t1), A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._eventData, _s10_, "eventData", t1), A.BuiltValueNullFieldError_checkNotNull(_this.get$_debug_event$_$this()._timestamp, _s10_, "timestamp", type$.int));
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.DebugEvent);
       return _this._debug_event$_$v = _$result;
     }
   };
@@ -22490,17 +22464,10 @@
       return _this;
     },
     _debug_event$_build$0() {
-      var _$failedField, e, _$result0, t1, exception, t2, _this = this,
-        _s18_ = "BatchedDebugEvents",
-        _$result = null;
+      var _$failedField, e, _$result0, exception, t1, _this = this, _$result = null;
       try {
         _$result0 = _this._debug_event$_$v;
-        if (_$result0 == null) {
-          t1 = _this.get$events().build$0();
-          _$result0 = new A._$BatchedDebugEvents(t1);
-          A.BuiltValueNullFieldError_checkNotNull(t1, _s18_, "events", type$.BuiltList_DebugEvent);
-        }
-        _$result = _$result0;
+        _$result = _$result0 == null ? new A._$BatchedDebugEvents(_this.get$events().build$0()) : _$result0;
       } catch (exception) {
         _$failedField = A._Cell$named("_$failedField");
         try {
@@ -22508,15 +22475,12 @@
           _this.get$events().build$0();
         } catch (exception) {
           e = A.unwrapException(exception);
-          t1 = A.BuiltValueNestedFieldError$(_s18_, _$failedField.readLocal$0(), J.toString$0$(e));
+          t1 = A.BuiltValueNestedFieldError$("BatchedDebugEvents", _$failedField.readLocal$0(), J.toString$0$(e));
           throw A.wrapException(t1);
         }
         throw exception;
       }
-      t1 = type$.BatchedDebugEvents;
-      t2 = t1._as(_$result);
-      A.ArgumentError_checkNotNull(t2, "other", t1);
-      _this._debug_event$_$v = t2;
+      _this._debug_event$_$v = type$.BatchedDebugEvents._as(_$result);
       return _$result;
     },
     set$_events(_events) {
@@ -22736,10 +22700,7 @@
     _build$0() {
       var _this = this,
         _$result = _this._$v;
-      if (_$result == null)
-        _$result = new A._$DebugInfo(_this.get$_$this()._appEntrypointPath, _this.get$_$this()._appId, _this.get$_$this()._appInstanceId, _this.get$_$this()._appOrigin, _this.get$_$this()._appUrl, _this.get$_$this()._authUrl, _this.get$_$this()._dwdsVersion, _this.get$_$this()._extensionUrl, _this.get$_$this()._isInternalBuild, _this.get$_$this()._isFlutterApp, _this.get$_$this()._workspaceName, _this.get$_$this()._tabUrl, _this.get$_$this()._tabId);
-      A.ArgumentError_checkNotNull(_$result, "other", type$.DebugInfo);
-      return _this._$v = _$result;
+      return _this._$v = _$result == null ? new A._$DebugInfo(_this.get$_$this()._appEntrypointPath, _this.get$_$this()._appId, _this.get$_$this()._appInstanceId, _this.get$_$this()._appOrigin, _this.get$_$this()._appUrl, _this.get$_$this()._authUrl, _this.get$_$this()._dwdsVersion, _this.get$_$this()._extensionUrl, _this.get$_$this()._isInternalBuild, _this.get$_$this()._isFlutterApp, _this.get$_$this()._workspaceName, _this.get$_$this()._tabUrl, _this.get$_$this()._tabId) : _$result;
     }
   };
   A.DevToolsRequest.prototype = {};
@@ -22845,8 +22806,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, _$result, t2, t3,
-        _s15_ = "promptExtension",
+      var t1, value, _$result,
         _s16_ = "DevToolsResponse",
         result = new A.DevToolsResponseBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -22878,13 +22838,8 @@
       _$result = result._devtools_request$_$v;
       if (_$result == null) {
         t1 = type$.bool;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_devtools_request$_$this()._success, _s16_, "success", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(result.get$_devtools_request$_$this()._promptExtension, _s16_, _s15_, t1);
-        _$result = new A._$DevToolsResponse(t2, t3, result.get$_devtools_request$_$this()._error);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s16_, "success", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s16_, _s15_, t1);
+        _$result = new A._$DevToolsResponse(A.BuiltValueNullFieldError_checkNotNull(result.get$_devtools_request$_$this()._success, _s16_, "success", t1), A.BuiltValueNullFieldError_checkNotNull(result.get$_devtools_request$_$this()._promptExtension, _s16_, "promptExtension", t1), result.get$_devtools_request$_$this()._error);
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.DevToolsResponse);
       return result._devtools_request$_$v = _$result;
     },
     deserialize$2(serializers, serialized) {
@@ -22941,19 +22896,13 @@
       return _this;
     },
     _devtools_request$_build$0() {
-      var t1, t2, t3, _this = this,
+      var t1, _this = this,
         _s15_ = "DevToolsRequest",
-        _s10_ = "instanceId",
         _$result = _this._devtools_request$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_devtools_request$_$this()._devtools_request$_appId, _s15_, "appId", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_devtools_request$_$this()._devtools_request$_instanceId, _s15_, _s10_, t1);
-        _$result = new A._$DevToolsRequest(t2, t3, _this.get$_devtools_request$_$this()._contextId, _this.get$_devtools_request$_$this()._devtools_request$_tabUrl, _this.get$_devtools_request$_$this()._uriOnly, _this.get$_devtools_request$_$this()._devtools_request$_client);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s15_, "appId", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s15_, _s10_, t1);
+        _$result = new A._$DevToolsRequest(A.BuiltValueNullFieldError_checkNotNull(_this.get$_devtools_request$_$this()._devtools_request$_appId, _s15_, "appId", t1), A.BuiltValueNullFieldError_checkNotNull(_this.get$_devtools_request$_$this()._devtools_request$_instanceId, _s15_, "instanceId", t1), _this.get$_devtools_request$_$this()._contextId, _this.get$_devtools_request$_$this()._devtools_request$_tabUrl, _this.get$_devtools_request$_$this()._uriOnly, _this.get$_devtools_request$_$this()._devtools_request$_client);
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.DevToolsRequest);
       return _this._devtools_request$_$v = _$result;
     }
   };
@@ -23004,8 +22953,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, $$v, _$result, t2, t3,
-        _s10_ = "stackTrace",
+      var t1, value, $$v, _$result,
         _s13_ = "ErrorResponse",
         result = new A.ErrorResponseBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -23045,13 +22993,8 @@
       _$result = result._error_response$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_error_response$_$this()._error_response$_error, _s13_, "error", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(result.get$_error_response$_$this()._error_response$_stackTrace, _s13_, _s10_, t1);
-        _$result = new A._$ErrorResponse(t2, t3);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s13_, "error", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s13_, _s10_, t1);
+        _$result = new A._$ErrorResponse(A.BuiltValueNullFieldError_checkNotNull(result.get$_error_response$_$this()._error_response$_error, _s13_, "error", t1), A.BuiltValueNullFieldError_checkNotNull(result.get$_error_response$_$this()._error_response$_stackTrace, _s13_, "stackTrace", t1));
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ErrorResponse);
       return result._error_response$_$v = _$result;
     },
     deserialize$2(serializers, serialized) {
@@ -23117,7 +23060,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, _$result, t2, t3, t4,
+      var t1, value, _$result,
         _s16_ = "ExtensionRequest",
         result = new A.ExtensionRequestBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -23147,17 +23090,7 @@
         }
       }
       _$result = result._extension_request$_$v;
-      if (_$result == null) {
-        t1 = type$.int;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._id, _s16_, "id", t1);
-        t3 = type$.String;
-        t4 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._command, _s16_, "command", t3);
-        _$result = new A._$ExtensionRequest(t2, t4, result.get$_extension_request$_$this()._commandParams);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s16_, "id", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s16_, "command", t3);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ExtensionRequest);
-      return result._extension_request$_$v = _$result;
+      return result._extension_request$_$v = _$result == null ? new A._$ExtensionRequest(A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._id, _s16_, "id", type$.int), A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._command, _s16_, "command", type$.String), result.get$_extension_request$_$this()._commandParams) : _$result;
     },
     deserialize$2(serializers, serialized) {
       return this.deserialize$3$specifiedType(serializers, serialized, B.FullType_null_List_empty_false);
@@ -23187,7 +23120,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, _$result, t2, t3, t4, t5, t6,
+      var t1, value, _$result,
         _s17_ = "ExtensionResponse",
         result = new A.ExtensionResponseBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -23223,20 +23156,7 @@
         }
       }
       _$result = result._extension_request$_$v;
-      if (_$result == null) {
-        t1 = type$.int;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._id, _s17_, "id", t1);
-        t3 = type$.bool;
-        t4 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_success, _s17_, "success", t3);
-        t5 = type$.String;
-        t6 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_result, _s17_, "result", t5);
-        _$result = new A._$ExtensionResponse(t2, t4, t6, result.get$_extension_request$_$this()._extension_request$_error);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s17_, "id", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s17_, "success", t3);
-        A.BuiltValueNullFieldError_checkNotNull(t6, _s17_, "result", t5);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ExtensionResponse);
-      return result._extension_request$_$v = _$result;
+      return result._extension_request$_$v = _$result == null ? new A._$ExtensionResponse(A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._id, _s17_, "id", type$.int), A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_success, _s17_, "success", type$.bool), A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_result, _s17_, "result", type$.String), result.get$_extension_request$_$this()._extension_request$_error) : _$result;
     },
     deserialize$2(serializers, serialized) {
       return this.deserialize$3$specifiedType(serializers, serialized, B.FullType_null_List_empty_false);
@@ -23259,7 +23179,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, $$v, _$result, t2, t3,
+      var t1, value, $$v, _$result,
         _s14_ = "ExtensionEvent",
         result = new A.ExtensionEventBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -23299,13 +23219,8 @@
       _$result = result._extension_request$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._params, _s14_, "params", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_method, _s14_, "method", t1);
-        _$result = new A._$ExtensionEvent(t2, t3);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s14_, "params", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s14_, "method", t1);
+        _$result = new A._$ExtensionEvent(A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._params, _s14_, "params", t1), A.BuiltValueNullFieldError_checkNotNull(result.get$_extension_request$_$this()._extension_request$_method, _s14_, "method", t1));
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ExtensionEvent);
       return result._extension_request$_$v = _$result;
     },
     deserialize$2(serializers, serialized) {
@@ -23541,17 +23456,10 @@
       return t1;
     },
     _extension_request$_build$0() {
-      var _$failedField, e, _$result0, t1, exception, t2, _this = this,
-        _s13_ = "BatchedEvents",
-        _$result = null;
+      var _$failedField, e, _$result0, exception, t1, _this = this, _$result = null;
       try {
         _$result0 = _this._extension_request$_$v;
-        if (_$result0 == null) {
-          t1 = _this.get$events().build$0();
-          _$result0 = new A._$BatchedEvents(t1);
-          A.BuiltValueNullFieldError_checkNotNull(t1, _s13_, "events", type$.BuiltList_ExtensionEvent);
-        }
-        _$result = _$result0;
+        _$result = _$result0 == null ? new A._$BatchedEvents(_this.get$events().build$0()) : _$result0;
       } catch (exception) {
         _$failedField = A._Cell$named("_$failedField");
         try {
@@ -23559,15 +23467,12 @@
           _this.get$events().build$0();
         } catch (exception) {
           e = A.unwrapException(exception);
-          t1 = A.BuiltValueNestedFieldError$(_s13_, _$failedField.readLocal$0(), J.toString$0$(e));
+          t1 = A.BuiltValueNestedFieldError$("BatchedEvents", _$failedField.readLocal$0(), J.toString$0$(e));
           throw A.wrapException(t1);
         }
         throw exception;
       }
-      t1 = type$.BatchedEvents;
-      t2 = t1._as(_$result);
-      A.ArgumentError_checkNotNull(t2, "other", t1);
-      _this._extension_request$_$v = t2;
+      _this._extension_request$_$v = type$.BatchedEvents._as(_$result);
       return _$result;
     },
     set$_extension_request$_events(_events) {
@@ -23583,8 +23488,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, $$v, _$result, t2,
-        _s16_ = "HotReloadRequest",
+      var t1, value, $$v, _$result,
         result = new A.HotReloadRequestBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
       for (; iterator.moveNext$0();) {
@@ -23608,14 +23512,7 @@
         }
       }
       _$result = result._hot_reload_request$_$v;
-      if (_$result == null) {
-        t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_hot_reload_request$_$this()._hot_reload_request$_id, _s16_, "id", t1);
-        _$result = new A._$HotReloadRequest(t2);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s16_, "id", t1);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.HotReloadRequest);
-      return result._hot_reload_request$_$v = _$result;
+      return result._hot_reload_request$_$v = _$result == null ? new A._$HotReloadRequest(A.BuiltValueNullFieldError_checkNotNull(result.get$_hot_reload_request$_$this()._hot_reload_request$_id, "HotReloadRequest", "id", type$.String)) : _$result;
     },
     deserialize$2(serializers, serialized) {
       return this.deserialize$3$specifiedType(serializers, serialized, B.FullType_null_List_empty_false);
@@ -23763,20 +23660,10 @@
       return _this;
     },
     _hot_reload_response$_build$0() {
-      var t1, t2, t3, t4, _this = this,
+      var _this = this,
         _s17_ = "HotReloadResponse",
         _$result = _this._hot_reload_response$_$v;
-      if (_$result == null) {
-        t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_hot_reload_response$_$this()._hot_reload_response$_id, _s17_, "id", t1);
-        t3 = type$.bool;
-        t4 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_hot_reload_response$_$this()._hot_reload_response$_success, _s17_, "success", t3);
-        _$result = new A._$HotReloadResponse(t2, t4, _this.get$_hot_reload_response$_$this()._errorMessage);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s17_, "id", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s17_, "success", t3);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.HotReloadResponse);
-      return _this._hot_reload_response$_$v = _$result;
+      return _this._hot_reload_response$_$v = _$result == null ? new A._$HotReloadResponse(A.BuiltValueNullFieldError_checkNotNull(_this.get$_hot_reload_response$_$this()._hot_reload_response$_id, _s17_, "id", type$.String), A.BuiltValueNullFieldError_checkNotNull(_this.get$_hot_reload_response$_$this()._hot_reload_response$_success, _s17_, "success", type$.bool), _this.get$_hot_reload_response$_$this()._errorMessage) : _$result;
     }
   };
   A.HotRestartRequest.prototype = {};
@@ -24034,10 +23921,7 @@
   A.IsolateExitBuilder.prototype = {
     _isolate_events$_build$0() {
       var _$result = this._isolate_events$_$v;
-      if (_$result == null)
-        _$result = new A._$IsolateExit();
-      A.ArgumentError_checkNotNull(_$result, "other", type$.IsolateExit);
-      return this._isolate_events$_$v = _$result;
+      return this._isolate_events$_$v = _$result == null ? new A._$IsolateExit() : _$result;
     }
   };
   A._$IsolateStart.prototype = {
@@ -24058,10 +23942,7 @@
   A.IsolateStartBuilder.prototype = {
     _isolate_events$_build$0() {
       var _$result = this._isolate_events$_$v;
-      if (_$result == null)
-        _$result = new A._$IsolateStart();
-      A.ArgumentError_checkNotNull(_$result, "other", type$.IsolateStart);
-      return this._isolate_events$_$v = _$result;
+      return this._isolate_events$_$v = _$result == null ? new A._$IsolateStart() : _$result;
     }
   };
   A.RegisterEvent.prototype = {};
@@ -24155,22 +24036,10 @@
       return _this;
     },
     _register_event$_build$0() {
-      var t1, t2, t3, t4, _this = this,
+      var _this = this,
         _s13_ = "RegisterEvent",
-        _s9_ = "eventData",
-        _s9_0 = "timestamp",
         _$result = _this._register_event$_$v;
-      if (_$result == null) {
-        t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_register_event$_$this()._register_event$_eventData, _s13_, _s9_, t1);
-        t3 = type$.int;
-        t4 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_register_event$_$this()._register_event$_timestamp, _s13_, _s9_0, t3);
-        _$result = new A._$RegisterEvent(t2, t4);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s13_, _s9_, t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s13_, _s9_0, t3);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.RegisterEvent);
-      return _this._register_event$_$v = _$result;
+      return _this._register_event$_$v = _$result == null ? new A._$RegisterEvent(A.BuiltValueNullFieldError_checkNotNull(_this.get$_register_event$_$this()._register_event$_eventData, _s13_, "eventData", type$.String), A.BuiltValueNullFieldError_checkNotNull(_this.get$_register_event$_$this()._register_event$_timestamp, _s13_, "timestamp", type$.int)) : _$result;
     }
   };
   A.RunRequest.prototype = {};
@@ -24183,11 +24052,8 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var _$result;
       type$.Iterable_nullable_Object._as(serialized);
-      _$result = new A._$RunRequest();
-      A.ArgumentError_checkNotNull(_$result, "other", type$.RunRequest);
-      return _$result;
+      return new A._$RunRequest();
     },
     deserialize$2(serializers, serialized) {
       return this.deserialize$3$specifiedType(serializers, serialized, B.FullType_null_List_empty_false);
@@ -24238,8 +24104,7 @@
       return this.serialize$3$specifiedType(serializers, object, B.FullType_null_List_empty_false);
     },
     deserialize$3$specifiedType(serializers, serialized, specifiedType) {
-      var t1, value, _$result, t2, t3, t4,
-        _s8_ = "argsJson",
+      var t1, value, _$result,
         _s23_ = "ServiceExtensionRequest",
         result = new A.ServiceExtensionRequestBuilder(),
         iterator = J.get$iterator$ax(type$.Iterable_nullable_Object._as(serialized));
@@ -24273,15 +24138,8 @@
       _$result = result._service_extension_request$_$v;
       if (_$result == null) {
         t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._service_extension_request$_id, _s23_, "id", t1);
-        t3 = A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._service_extension_request$_method, _s23_, "method", t1);
-        t4 = A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._argsJson, _s23_, _s8_, t1);
-        _$result = new A._$ServiceExtensionRequest(t2, t3, t4);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s23_, "id", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t3, _s23_, "method", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t4, _s23_, _s8_, t1);
+        _$result = new A._$ServiceExtensionRequest(A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._service_extension_request$_id, _s23_, "id", t1), A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._service_extension_request$_method, _s23_, "method", t1), A.BuiltValueNullFieldError_checkNotNull(result.get$_service_extension_request$_$this()._argsJson, _s23_, "argsJson", t1));
       }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ServiceExtensionRequest);
       return result._service_extension_request$_$v = _$result;
     },
     deserialize$2(serializers, serialized) {
@@ -24473,21 +24331,10 @@
       return _this;
     },
     _service_extension_response$_build$0() {
-      var t1, t2, t3, t4, t5, _this = this,
+      var _this = this,
         _s24_ = "ServiceExtensionResponse",
         _$result = _this._service_extension_response$_$v;
-      if (_$result == null) {
-        t1 = type$.String;
-        t2 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_service_extension_response$_$this()._service_extension_response$_id, _s24_, "id", t1);
-        t3 = _this.get$_service_extension_response$_$this()._resultJson;
-        t4 = type$.bool;
-        t5 = A.BuiltValueNullFieldError_checkNotNull(_this.get$_service_extension_response$_$this()._service_extension_response$_success, _s24_, "success", t4);
-        _$result = new A._$ServiceExtensionResponse(t2, t3, t5, _this.get$_service_extension_response$_$this()._errorCode, _this.get$_service_extension_response$_$this()._service_extension_response$_errorMessage);
-        A.BuiltValueNullFieldError_checkNotNull(t2, _s24_, "id", t1);
-        A.BuiltValueNullFieldError_checkNotNull(t5, _s24_, "success", t4);
-      }
-      A.ArgumentError_checkNotNull(_$result, "other", type$.ServiceExtensionResponse);
-      return _this._service_extension_response$_$v = _$result;
+      return _this._service_extension_response$_$v = _$result == null ? new A._$ServiceExtensionResponse(A.BuiltValueNullFieldError_checkNotNull(_this.get$_service_extension_response$_$this()._service_extension_response$_id, _s24_, "id", type$.String), _this.get$_service_extension_response$_$this()._resultJson, A.BuiltValueNullFieldError_checkNotNull(_this.get$_service_extension_response$_$this()._service_extension_response$_success, _s24_, "success", type$.bool), _this.get$_service_extension_response$_$this()._errorCode, _this.get$_service_extension_response$_$this()._service_extension_response$_errorMessage) : _$result;
     }
   };
   A.BatchedStreamController.prototype = {
@@ -27906,7 +27753,7 @@
             case 8:
               // then
               $async$goto = 11;
-              return A._asyncAwait($async$self.manager.hotRestart$0(), $async$call$1);
+              return A._asyncAwait($async$self.manager.hotRestart$1$reloadedSourcesPath(A._asStringQ(t1.$reloadedSourcesPath)), $async$call$1);
             case 11:
               // returning from await.
               // goto join
@@ -28466,11 +28313,8 @@
       });
       return A._asyncStartSync($async$hotRestart$3$readyToRunMain$reloadedSourcesPath$runId, $async$completer);
     },
-    hotRestart$0() {
-      return this.hotRestart$3$readyToRunMain$reloadedSourcesPath$runId(null, null, null);
-    },
-    hotRestart$1$runId(runId) {
-      return this.hotRestart$3$readyToRunMain$reloadedSourcesPath$runId(null, null, runId);
+    hotRestart$1$reloadedSourcesPath(reloadedSourcesPath) {
+      return this.hotRestart$3$readyToRunMain$reloadedSourcesPath$runId(null, reloadedSourcesPath, null);
     },
     hotRestart$2$reloadedSourcesPath$runId(reloadedSourcesPath, runId) {
       return this.hotRestart$3$readyToRunMain$reloadedSourcesPath$runId(null, reloadedSourcesPath, runId);
@@ -29279,8 +29123,6 @@
       BuildResult: findType("BuildResult"),
       BuildStatus: findType("BuildStatus"),
       BuiltListMultimap_dynamic_dynamic: findType("BuiltListMultimap<@,@>"),
-      BuiltList_DebugEvent: findType("BuiltList<DebugEvent>"),
-      BuiltList_ExtensionEvent: findType("BuiltList<ExtensionEvent>"),
       BuiltList_dynamic: findType("BuiltList<@>"),
       BuiltList_nullable_Object: findType("BuiltList<Object?>"),
       BuiltMap_dynamic_dynamic: findType("BuiltMap<@,@>"),