+++ dwds/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 25.0.2
+
+### Bug Fixes:
+
+- Fix issue in hot restart with the web socket where we didn't pass the reloaded
+  sources path, resulting in a null assertion.
+
 ## 25.0.1
 
 ### Bug Fixes: